LayoutTests/media/wireless-playback-media-player/load-url-source-element-fallback-2.html

<!DOCTYPE html><!-- webkit-test-runner [ WirelessPlaybackMediaPlayerEnabled=true ] -->
<html>
    <head>
        <script src='../media-file.js'></script>
        <script src='../video-test.js'></script>
        <script>
            var route;
            var video;
            var actualURL;

            async function start()
            {
                if (window.internals)
                    internals.mockMediaDeviceRouteController.enabled = true;

                video = document.createElement('video');

                const mediaType = 'video/mp4';
                const mediaURL = findMediaFile('video', '../content/test');
                const mediaURLToReject = findMediaFile('video', '../content/test_yuv420');

                const firstSource = document.createElement('source');
                firstSource.src = mediaURLToReject;
                firstSource.type = mediaType;
                video.appendChild(firstSource);

                const secondSource = document.createElement('source');
                secondSource.src = mediaURL;
                secondSource.type = mediaType;
                video.appendChild(secondSource);

                document.body.appendChild(video);

                await waitFor(video, 'canplaythrough');

                run(`video.play()`)
                await waitFor(video, 'playing');

                route = internals.mockMediaDeviceRouteController.createMockMediaDeviceRoute();

                const firstURLPromise = new Promise((resolve) => {
                    route.setURLCallback(async (url) => {
                        resolve(url);
                    });
                });

                run(`internals.mockMediaDeviceRouteController.activateRoute(route)`);
                await firstURLPromise;

                const secondURLPromise = new Promise((resolve) => {
                    route.setURLCallback(async (url) => {
                        resolve(url);
                    });
                });

                run(`route.hasPlaybackError = true`);

                const loadedURL = await secondURLPromise;
                const expectedURL = new URL(mediaURL, document.baseURI).href.substring(document.baseURI.indexOf('/LayoutTests/'));
                actualURL = loadedURL.substring(loadedURL.indexOf('/LayoutTests/'));

                testExpected('actualURL', expectedURL);

                endTest();
            }
        </script>
    </head>
    <body onload='start()'>
        <p>Test that a video's second &lt;source&gt; element's URL is loaded if the MediaDeviceRoute rejects encounters a playback error loading the first URL prior to becoming ready.</p>
    </body>
</html>
